refactor: split the designer share validator into per-area modules
api/lib/designerValidation.ts had 1431 code lines. The colour, text, cutout,
lid and base validators move to sibling modules alongside the existing
compartment validator, each with the enum tables it reads. Dependencies run
one way (cutouts on text and colour, lid on cutouts, base on lid) and the
orchestrator keeps exporting every name it exported before.
